+++ R/find-nonmem.R
@@ -1,5 +1,107 @@
 psn_available <- function() Sys.which("psn") != ""
 
+#' Tests if psn is accessible
+#' 
+#' @description
+#'
+#' `r lifecycle::badge("experimental")`
+#' 
+#' NMproject needs PsN installed to function.  If PsN is not installed.
+#' 
+#' @return Logical TRUE or FALSE.
+#' 
+#' @examples 
+#' 
+#' psn_available()
+#' 
+#' @keywords internal
+#' @export
+
+psn_check <- function() {
+  
+  system_nm_available(fail_if_false = TRUE)
+  
+  ## check psn
+  
+  psn_nm_versions <- try(system_nm_intern("psn --version"), silent = TRUE)
+  
+  if (inherits(psn_nm_versions, "try-error")) {
+    if (fail_if_false) {
+      usethis::ui_oops("{usethis::ui_code('system_nm(\"psn --version\")')} command fails.  Ensure PsN is installed: {usethis::ui_path('https://uupharmacometrics.github.io/PsN/')}")
+    }
+    return(FALSE)
+  }
+  
+  if (length(psn_nm_versions) == 0) {
+    if (fail_if_false) {
+      usethis::ui_oops("{usethis::ui_code('system_nm(\"psn --version\")')} command fails to return a result.  Ensure PsN is installed: {usethis::ui_path('https://uupharmacometrics.github.io/PsN/')}")
+    }
+    return(FALSE)
+  }
+  
+  return(TRUE)
+  
+}
+
+#' Test integrity of system_nm
+#'
+#' @description
+#'
+#' `r lifecycle::badge("stable")`
+#'
+#' `system_nm()` needs to be able to submit commands to the terminal for
+#' NMproject to function.  This function tests whether meaningful output can be
+#' produced and is only intended for diagnostic purposes.
+#'
+#' @param fail_if_false Logical (default = `FALSE`). Should the function return
+#'   an error message if `system_nm()` is not available.
+#'
+#' @return Logical TRUE or FALSE.
+#'
+#' @examples
+#'
+#' system_nm_available()
+#'
+#' @keywords internal
+#' @export
+
+system_nm_available <- function(fail_if_false = FALSE) {
+  result <- try(system_nm_intern("whoami"), silent = TRUE)
+  
+  if (inherits(result, "try-error")) {
+    if (fail_if_false) {
+      usethis::ui_oops("{usethis::ui_code('system_nm(\"whoami\")')} command fails.  See {usethis::ui_code('?system_nm')} for more information how to configure")
+    }
+    return(FALSE)
+  }
+  
+  if (length(result) == 0) {
+    if (fail_if_false) {
+      usethis::ui_oops("{usethis::ui_code('system_nm(\"whoami\")')} command fails to return a result.  See {usethis::ui_code('?system_nm')} for more information how to configure")
+    }
+    return(FALSE)
+  }
+  
+  result <- result[1]
+  
+  if (!is.character(result)) {
+    if (fail_if_false) {
+      usethis::ui_oops("{usethis::ui_code('system_nm(\"whoami\")')} command fails to return a character type.  See {usethis::ui_code('?system_nm')} for more information how to configure")
+    }
+    return(FALSE)
+  }
+  
+  if (nchar(result) == 0) {
+    if (fail_if_false) {
+      usethis::ui_oops("{usethis::ui_code('system_nm(\"whoami\")')} command fails to return a meaningful output.  See {usethis::ui_code('?system_nm')} for more information how to configure")
+    }
+    return(FALSE)
+  }
+  
+  return(TRUE)
+
+}
+
 #' @rdname find_nonmem
 #' @name find_nonmem
 #' @title Find location of NONMEM installation
@@ -13,9 +115,10 @@ psn_available <- function() Sys.which("psn") != ""
 #'
 #' @param name Character name of NONMEM installation (according PsN).
 #'
-#' @details The function `find_nm_install_path()` will attempt to use a locally available PsN
-#'   installation to get this information.  If the PsN installation is on a
-#'   remote server, this function will not work (it will return a `NULL`)
+#' @details The function `find_nm_install_path()` will attempt to use a locally
+#'   available PsN installation to get this information.  If the PsN
+#'   installation is on a remote server, this function will not work (it will
+#'   return a `NULL`)
 #' @return If functions cannot find installation they will return `NULL` without
 #'   errors or warning, otherwise they will return the located paths.
 #'
@@ -25,9 +128,8 @@ find_nm_install_path <- function(name = "default") {
     return(NULL)
   }
 
-  nm_versions <- try(system_nm("psn -nm_versions", intern = TRUE, ignore.stderr = TRUE, wait = TRUE),
-    silent = TRUE
-  )
+  nm_versions <- try(system_nm_intern("psn -nm_versions"), silent = TRUE)
+  
   if (inherits(nm_versions, "try-error")) {
     return(NULL)
   }
@@ -62,9 +164,7 @@ find_nm_tran_path <- function(name = "default", warn = TRUE) {
     return(NULL)
   }
 
-  nm_versions <- try(system_nm("psn -nm_versions", intern = TRUE, ignore.stderr = TRUE, wait = TRUE),
-    silent = TRUE
-  )
+  nm_versions <- try(system_nm_intern("psn -nm_versions"), silent = TRUE)
 
   warn_func <- function() {
     if (warn) {
@@ -119,9 +219,7 @@ find_nm_tran_path <- function(name = "default", warn = TRUE) {
 NONMEM_version <- function() {
 
   ## scrape version info
-  psn_nm_versions <- try(system_nm("psn -nm_versions", intern = TRUE, ignore.stderr = TRUE, wait = TRUE),
-    silent = TRUE
-  )
+  psn_nm_versions <- try(system_nm_intern("psn -nm_versions"), silent = TRUE)
   if (inherits(psn_nm_versions, "try-error")) {
     warning("can't find nonmem installation")
     return(NULL)
@@ -132,15 +230,15 @@ NONMEM_version <- function() {
 
   if (length(psn_nm_version) == 0) psn_nm_version <- "not found"
 
-  nm_compiler_version <- system_nm("gfortran --version", intern = TRUE, ignore.stderr = TRUE, wait = TRUE)[1]
+  nm_compiler_version <- system_nm_intern("gfortran --version")[1]
   if (length(nm_compiler_version) == 0) {
     nm_compiler_version <- "not found"
   }
 
-  psn_version <- system_nm("psn --version", intern = TRUE, ignore.stderr = TRUE, wait = TRUE)
+  psn_version <- system_nm_intern("psn --version")
   if (length(psn_version) == 0) psn_version <- "not found"
 
-  perl_version <- system_nm("perl --version", intern = TRUE, ignore.stderr = TRUE, wait = TRUE)
+  perl_version <- system_nm_intern("perl --version")
   perl_version <- perl_version[grepl("This is", perl_version)]
   perl_version <- gsub("^.*\\((v.*)\\).*$", "\\1", perl_version)
   if (length(perl_version) == 0) perl_version <- "not found"